+++ packages/blade/src/components/Button/Button/Button.tsx
@@ -1,6 +1,7 @@
 import type { GestureResponderEvent } from 'react-native';
 import React from 'react';
 import BaseButton from '../BaseButton';
+import type { BaseButtonProps } from '../BaseButton/BaseButton';
 import type { IconComponent } from '~components/Icons';
 import type { Platform } from '~utils';
 import type { StyledPropsBlade } from '~components/Box/styledProps';
@@ -10,6 +11,22 @@ import { assignWithoutSideEffects } from '~src/utils/assignWithoutSideEffects';
 import type { TooltipTriggerProps } from '~components/Tooltip/types';
 
 type ButtonCommonProps = {
+  /**
+   * Automatically renders button with `a` tag with `href` on web
+   */
+  href?: BaseButtonProps['href'];
+  /**
+   * anchor target attribute
+   *
+   * Should only be used alongside `href`
+   */
+  target?: BaseButtonProps['target'];
+  /**
+   * anchor rel attribute
+   *
+   * Should only be used alongside `href`
+   */
+  rel?: BaseButtonProps['rel'];
   variant?: 'primary' | 'secondary' | 'tertiary';
   size?: 'xsmall' | 'small' | 'medium' | 'large';
   iconPosition?: 'left' | 'right';
@@ -53,6 +70,9 @@ const _Button: React.ForwardRefRenderFunction<BladeElementRef, ButtonProps> = (
     isDisabled = false,
     isFullWidth = false,
     isLoading = false,
+    href,
+    target,
+    rel,
     onClick,
     size = 'medium',
     type = 'button',
@@ -74,6 +94,9 @@ const _Button: React.ForwardRefRenderFunction<BladeElementRef, ButtonProps> = (
       {...(icon ? { icon, children } : { children })}
       {...styledProps}
       ref={ref}
+      href={href}
+      target={target}
+      rel={rel}
       accessibilityLabel={accessibilityLabel}
       iconPosition={iconPosition}
       isDisabled={isDisabled}