+++ packages/blade/src/components/Tooltip/Tooltip.native.tsx
@@ -1,91 +1,76 @@
 /* eslint-disable @typescript-eslint/restrict-plus-operands */
-import type { Side } from '@floating-ui/react-native';
 import { arrow, shift, useFloating, flip, offset } from '@floating-ui/react-native';
 import React from 'react';
-import { Modal, Pressable, TouchableOpacity } from 'react-native';
+import { Modal, TouchableOpacity } from 'react-native';
 import { TooltipArrow } from './TooltipArrowNative';
 import { TooltipContent } from './TooltipContent';
 import { TooltipProps } from './types';
 import { ARROW_HEIGHT, ARROW_WIDTH } from './constants';
+import { getPlacementParts } from './utils';
 import { useTheme } from '~components/BladeProvider';
 
 const Tooltip = ({
   content,
   children,
-  onClose,
-  onOpen,
   placement = 'left',
-  shouldWrapChildren,
+  onOpenChange,
 }: TooltipProps): React.ReactElement => {
   const { theme } = useTheme();
   const [isOpen, setIsOpen] = React.useState(false);
 
   const gap = theme.spacing[2];
-  const [side] = placement.split('-') as [Side];
+  const [side] = getPlacementParts(placement);
   const isHorizontal = side === 'left' || side === 'right';
   const arrowRef = React.useRef();
   const context = useFloating({
     sameScrollView: false,
-    middleware: [
-      flip(),
-      shift({ padding: gap }),
-      offset(gap + ARROW_HEIGHT),
-      arrow({ element: arrowRef, padding: isHorizontal ? 0 : ARROW_WIDTH }),
-    ],
     placement,
+    middleware: [
+      shift({ crossAxis: false, padding: gap }),
+      flip({ padding: gap }),
+      offset(gap + ARROW_HEIGHT),
+      arrow({
+        element: arrowRef,
+        padding: isHorizontal ? 0 : ARROW_WIDTH,
+      }),
+    ],
   });
 
   const { refs, floatingStyles } = context;
 
   const handleOpen = React.useCallback(() => {
     setIsOpen(true);
-    onOpen?.();
-  }, [onOpen]);
+    onOpenChange?.({ isOpen: true });
+  }, [onOpenChange]);
 
   const handleClose = React.useCallback(() => {
     setIsOpen(false);
-    onClose?.();
-  }, [onClose]);
+    onOpenChange?.({ isOpen: false });
+  }, [onOpenChange]);
 
-  // TODO: Do we need shouldWrapChildren in ReactNative?
-  // We won't need to add shouldWrapChildren in RN, We will always wrap the children in a Pressable
-  // Because even if we support direct childrens the tooltip won't behave as expected
-  // In case if a interactive element is passed into it, Since we don't support long press event.
-  // This is inline with design: In mobile if we put tooltip around a interactive element like button
-  // that will be a wrong UX
-  // if (shouldWrapChildren) {
-  //   console.warn('[Blade Tooltip]: `shouldWrapChildren` prop does nothing on ReactNative');
-  // }
+  // wait for animation to finish before unmounting modal
+  const [isVisible, setIsVisible] = React.useState(() => isOpen);
+  React.useEffect(() => {
+    const id = setTimeout(() => {
+      if (!isOpen) {
+        setIsVisible(false);
+      }
+    }, theme.motion.duration.gentle);
+
+    if (isOpen) {
+      setIsVisible(true);
+    }
+    return () => clearTimeout(id);
+  }, [isOpen]);
 
   return (
     <>
-      {shouldWrapChildren ? (
-        <Pressable
-          style={{ alignSelf: 'flex-start' }}
-          // using touch end instead of start so that if the the children is interactive
-          // it's events will get triggered also
-          onTouchEnd={handleOpen}
-          ref={refs.setReference}
-          collapsable={false}
-          testID="tooltip-interactive-wrapper"
-        >
-          {children}
-        </Pressable>
-      ) : (
-        React.cloneElement(children, {
-          onTouchEnd: handleOpen,
-          ref: refs.setReference,
-          style: { alignSelf: 'flex-start' },
-        })
-      )}
-
-      <Modal
-        accessibilityLabel={content}
-        collapsable={false}
-        transparent
-        visible={isOpen}
-        animationType="fade"
-      >
+      {/* Cloning the trigger children, so enhance it with ref and event handler */}
+      {React.cloneElement(children, {
+        onTouchEnd: handleOpen,
+        ref: refs.setReference,
+      })}
+      <Modal collapsable={false} transparent visible={isVisible}>
         <TouchableOpacity
           style={{
             flexShrink: 0,
@@ -96,8 +81,17 @@ const Tooltip = ({
           testID="tooltip-modal-backdrop"
         >
           <TooltipContent
+            isVisible={isOpen}
             ref={refs.setFloating}
-            style={floatingStyles}
+            side={side}
+            style={{
+              ...floatingStyles,
+              // To avoid flash of floating ui content at top, this only happens in RN <70
+              // if the position is zero move the floating element outside of the viewport
+              // this happens because measure is async and it takes few miliseconds to calculate the positions.
+              left: floatingStyles.left || -200,
+              top: floatingStyles.top || -200,
+            }}
             arrow={<TooltipArrow context={context} ref={arrowRef as never} />}
           >
             {content}