+++ packages/blade/src/components/Tooltip/types.ts
@@ -1,6 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import type { UseFloatingOptions } from '@floating-ui/react';
+import type { Side, UseFloatingOptions } from '@floating-ui/react';
 import type { GestureResponderEvent, NativeSyntheticEvent } from 'react-native';
+import type { CSSProperties } from 'react';
 import type { Platform } from '~utils';
 
 type TooltipProps = {
@@ -10,9 +11,7 @@ type TooltipProps = {
     'left-end' | 'left-start' | 'right-end' | 'right-start'
   >;
   children: React.ReactElement;
-  shouldWrapChildren?: boolean;
-  onOpen?: () => void;
-  onClose?: () => void;
+  onOpenChange?: ({ isOpen }: { isOpen: boolean }) => void;
 };
 
 type TooltipTriggerProps = {
@@ -38,4 +37,18 @@ type TooltipTriggerProps = {
   }>;
 };
 
-export { TooltipProps, TooltipTriggerProps };
+type TooltipContentProps = {
+  children: React.ReactNode;
+  style: CSSProperties;
+  arrow: React.ReactNode;
+  /**
+   * react-native only
+   */
+  isVisible?: boolean;
+  /**
+   * react-native only
+   */
+  side?: Side;
+};
+
+export { TooltipProps, TooltipTriggerProps, TooltipContentProps };