+++ packages/blade/src/components/Tooltip/utils.ts
@@ -1,0 +1,9 @@
+import type { Alignment, Placement, Side } from '@floating-ui/react';
+
+// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
+const getPlacementParts = (placement: NonNullable<Placement>) => {
+  const [side, alignment] = placement.split('-') as [Side, Alignment];
+  return [side, alignment] as const;
+};
+
+export { getPlacementParts };